requirements.txt
asgiref==3.2.10
attrs==19.3.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
distro==1.5.0
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
docker==4.4.0
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
jsonschema==3.2.0
more-itertools==8.2.0
packaging==20.3
paramiko==2.7.2
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==5.4.1
pytest-django==3.9.0
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
texttable==1.6.3
urllib3==1.25.9
wcwidth==0.1.9
websocket-client==0.57.0
yapf==0.30.0
zipp==3.1.0
